+++ src/fides/api/models/privacy_request/execution_log.py
@@ -24,7 +24,7 @@ EXECUTION_CHECKPOINTS = [
     CurrentStep.finalize_consent,
     CurrentStep.email_post_send,
     CurrentStep.post_webhooks,
-    CurrentStep.manual_finalization,
+    CurrentStep.finalization,
 ]
 
 COMPLETED_EXECUTION_LOG_STATUSES = [
