+++ tests/ops/api/v1/endpoints/privacy_request/test_privacy_request_endpoints.py
@@ -92,7 +92,6 @@ from fides.common.api.v1.urn_registry import (
     PRIVACY_REQUEST_BULK_SOFT_DELETE,
     PRIVACY_REQUEST_DENY,
     PRIVACY_REQUEST_FILTERED_RESULTS,
-    PRIVACY_REQUEST_FINALIZE,
     PRIVACY_REQUEST_MANUAL_WEBHOOK_ACCESS_INPUT,
     PRIVACY_REQUEST_MANUAL_WEBHOOK_ERASURE_INPUT,
     PRIVACY_REQUEST_NOTIFICATIONS,
@@ -8757,74 +8756,3 @@ class TestSendBatchEmailIntegrations:
             == "Email batch job started. This may take a few minutes to complete."
         )
         mock_send_email_batch.assert_called_once()
-
-
-class TestFinalizePrivacyRequest:
-    @pytest.fixture(scope="function")
-    def url(self, privacy_request_requires_manual_finalization) -> str:
-        return V1_URL_PREFIX + PRIVACY_REQUEST_FINALIZE.format(
-            privacy_request_id=privacy_request_requires_manual_finalization.id
-        )
-
-    def test_finalize_privacy_request_unauthenticated(self, api_client, url):
-        response = api_client.post(url)
-        assert response.status_code == 401
-
-    def test_finalize_privacy_request_wrong_scope(
-        self, api_client, url, generate_auth_header
-    ):
-        auth_header = generate_auth_header(scopes=[PRIVACY_REQUEST_READ])
-        response = api_client.post(url, headers=auth_header)
-        assert response.status_code == 403
-
-    def test_finalize_privacy_request_wrong_status(
-        self, api_client, generate_auth_header, privacy_request
-    ):
-        url = V1_URL_PREFIX + PRIVACY_REQUEST_FINALIZE.format(
-            privacy_request_id=privacy_request.id
-        )
-        auth_header = generate_auth_header(scopes=[PRIVACY_REQUEST_REVIEW])
-        response = api_client.post(url, headers=auth_header)
-        assert response.status_code == 400
-        assert (
-            f"Privacy request with id '{privacy_request.id}' is not in the requires_manual_finalization state."
-            in response.json()["detail"]
-        )
-
-    @mock.patch(
-        "fides.api.api.v1.endpoints.privacy_request_endpoints.queue_privacy_request"
-    )
-    def test_finalize_privacy_request(
-        self,
-        mock_queue_privacy_request,
-        db,
-        api_client,
-        url,
-        generate_auth_header,
-        privacy_request_requires_manual_finalization,
-        erasure_request_finalization_required_true,
-        user,
-    ):
-        payload = {
-            JWE_PAYLOAD_ROLES: user.client.roles,
-            JWE_PAYLOAD_CLIENT_ID: user.client.id,
-            JWE_ISSUED_AT: datetime.now().isoformat(),
-        }
-        # User already has Privacy Request Review scope
-        auth_header = {
-            "Authorization": "Bearer "
-            + generate_jwe(json.dumps(payload), CONFIG.security.app_encryption_key)
-        }
-        response = api_client.post(url, headers=auth_header)
-        assert response.status_code == 200
-
-        privacy_request_requires_manual_finalization.refresh_from_db(db=db)
-        assert (
-            privacy_request_requires_manual_finalization.status
-            == PrivacyRequestStatus.requires_manual_finalization
-        )
-        assert privacy_request_requires_manual_finalization.finalized_by == user.id
-        assert privacy_request_requires_manual_finalization.finalized_at is not None
-        mock_queue_privacy_request.assert_called_with(
-            privacy_request_id=privacy_request_requires_manual_finalization.id
-        )
